Same for EnglishWordList (Electrum)
